Add explicit main project DLLs build to a separate docker layer to optimise image rebuild time.
It turns out that the whole main project is being built implicitly regardless of what classes are being consumed by the tests project. The tests project may as well use nothing from the main project, but so long as the 2 projects are linked through their csproj, if the main project wasn't built in full, the tests project build will complain. Seems build process is not optimised in dotnet.